pyre-object: pass str/bytes leaf operands as object pointers
`jit_str_find_objs`, `jit_str_rfind_objs`, `jit_str_count_objs`,
`jit_str_contains`, `jit_str_compare`, `jit_bytes_contains` and
`jit_bytes_contains_byte` took their `W_UnicodeObject` / `W_BytesObject`
and boxed-bound operands as `i64`, and their interpreter callers passed
`obj as i64`.  The codewriter lowers that cast to `cast_ptr_to_int`, so a
traced call carried the receiver as an Int box.  In the
`email.message._parseparam` loop the optimizer hoisted `CastPtrToInt(s)`
into the loop label, and after a minor collection moved `s` the int still
held the old nursery address: `test_email
test_get_param_linear_complexity` read a stale payload (`rutf8.rs` index
out of bounds / SIGSEGV; a wrong result under
`PYPY_GC_NURSERY=64K PYPY_GC_NURSERY_DEBUG=1`).

The string and bytes leaves now take `PyObjectRef`, so the residual call
records Ref arguments the GC map tracks.  The optimized `_parseparam`
trace has no `CastPtrToInt` left (6 before).

The `find` / `rfind` / `count` gateways passed an omitted bound as a
null object; as a Ref argument that null makes the walker refuse the
residual (`[nullref-refusal]`), which aborted `warm_find` in
`str_search_index_bounds`.  The gateways now unwrap the admitted bound
in the traced body (`str_search_bound_value`, `_convert_idx_params`) and
call `jit_str_find_bounds` / `jit_str_rfind_bounds` /
`jit_str_count_bounds`, the `ll_find(s1, s2, start, end)` shape; the
`_objs` leaves and `cp_bound_from_obj` are removed.

spec_folds! rows 72 -> 72, fn try_walker_specialize_ 65 -> 65.

+++ pyre/pyre-jit-trace/src/jitcode_dispatch/specialize.rs
@@ -20931,7 +20931,7 @@ pub(crate) fn try_walker_specialize_compare_op_str<Sym: WalkSym>(
     };
     // Read the ordering before the residual runs: both operands are exact
     // `str`, so nothing here can run Python code or move an object.
-    let ordering = pyre_object::unicodeobject::jit_str_compare(lhs_obj as i64, rhs_obj as i64);
+    let ordering = pyre_object::unicodeobject::jit_str_compare(lhs_obj, rhs_obj);
     let folded = majit_metainterp::eval_binop_i(cmp, ordering, 0);
 
     // The authentic answer, from the same may-force path the generic leg
